+++ tests/test_remote.py
@@ -6,16 +6,16 @@ import time
 import logging
 from plumbum import RemotePath, SshMachine, ProcessExecutionError, local, ProcessTimedOut, NOHUP
 from plumbum import CommandNotFound
-from plumbum.lib import six, ensure_skipIf
+from plumbum.lib import six
+from .unittesttools import skipIf, skip_without_chown
 
-ensure_skipIf(unittest)
 
 #TEST_HOST = "192.168.1.143"
 TEST_HOST = "127.0.0.1"
 if TEST_HOST not in ("::1", "127.0.0.1", "localhost"):
     import plumbum
     plumbum.local.env.path.append("c:\\Program Files\\Git\\bin")
-
+@skip_without_bash
 class RemotePathTest(unittest.TestCase):
     def _connect(self):
         return SshMachine(TEST_HOST)
@@ -51,7 +51,7 @@ class RemotePathTest(unittest.TestCase):
         strcmp(p1.with_name("something.tar"), RemotePath(self._connect(), "/some/long/path/to/something.tar"))
         strcmp(p2.with_name("something.tar"), RemotePath(self._connect(), "something.tar"))
 
-    @unittest.skipIf(not hasattr(os, "chown"), "os.chown not supported")
+    @skip_without_chown
     def test_chown(self):
         with self._connect() as rem:
             with rem.tempdir() as dir:
@@ -63,7 +63,7 @@ class RemotePathTest(unittest.TestCase):
                 p.chown(p.uid.name)
                 self.assertEqual(p.uid, os.getuid())
 
-
+@skip_without_bash
 class BaseRemoteMachineTest(object):
     TUNNEL_PROG = r"""import sys, socket
 s = socket.socket()
@@ -181,7 +181,7 @@ s.close()
             else:
                 self.fail("Expected an execution error")
 
-
+@skip_without_bash
 class RemoteMachineTest(unittest.TestCase, BaseRemoteMachineTest):
     def _connect(self):
         return SshMachine(TEST_HOST)
@@ -239,7 +239,7 @@ class RemoteMachineTest(unittest.TestCase, BaseRemoteMachineTest):
                 self.assertEqual(printenv.with_env(FOO = "sea", BAR = "world")("FOO"), "sea\n")
                 self.assertEqual(printenv.with_env(FOO = "sea", BAR = "world")("BAR"), "world\n")
 
-    @unittest.skipIf('useradd' not in local, "System does not have useradd (Mac?)")
+    @skipIf('useradd' not in local, "System does not have useradd (Mac?)")
     def test_sshpass(self):
         with local.as_root():
             local["useradd"]("-m", "-b", "/tmp", "testuser")
@@ -267,7 +267,7 @@ except ImportError:
     print("Paramiko not avilable")
 else:
     from plumbum.machines.paramiko_machine import ParamikoMachine
-
+    @skip_on_windows
     class TestParamikoMachine(unittest.TestCase, BaseRemoteMachineTest):
         def _connect(self):
             return ParamikoMachine(TEST_HOST, missing_host_policy = paramiko.AutoAddPolicy())
